+++ proxy/quic.go
@@ -34,7 +34,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"log"
 	"net"
 	"net/http"
 	"sync"
@@ -80,7 +79,7 @@ func (p *Proxy) startQUIC(ctx context.Context) error {
 				return be.tlsConfigQUIC, nil
 			}
 		}
-		log.Printf("ERR QUIC connection %s %s", hello.ServerName, hello.SupportedProtos)
+		p.logErrorF("ERR QUIC connection %s %s", hello.ServerName, hello.SupportedProtos)
 		return nil, tlsUnrecognizedName
 	}
 	qt, err := netw.NewQUIC(p.cfg.TLSAddr, statelessResetKey)
@@ -100,15 +99,15 @@ func (p *Proxy) startQUIC(ctx context.Context) error {
 }
 
 func (p *Proxy) quicAcceptLoop(ctx context.Context, ln *netw.QUICListener) {
-	log.Printf("INF Accepting QUIC connections on %s %s", ln.Addr().Network(), ln.Addr())
+	p.logConnF("INF Accepting QUIC connections on %s %s", ln.Addr().Network(), ln.Addr())
 	for {
 		conn, err := ln.Accept(ctx)
 		if err != nil {
 			if errors.Is(err, quic.ErrServerClosed) || errors.Is(err, context.Canceled) || err.Error() == "closing" {
-				log.Print("INF QUIC Accept loop terminated")
+				p.logErrorF("INF QUIC Accept loop terminated")
 				break
 			}
-			log.Printf("ERR QUIC Accept: %v", err)
+			p.logErrorF("ERR QUIC Accept: %v", err)
 			continue
 		}
 		go p.handleQUICConnection(conn)
@@ -119,7 +118,7 @@ func (p *Proxy) handleQUICConnection(qc *netw.QUICConn) {
 	defer func() {
 		if r := recover(); r != nil {
 			p.recordEvent("panic")
-			log.Printf("ERR [%s] %s: PANIC: %v", certSummary(connClientCert(qc)), qc.RemoteAddr(), r)
+			p.logErrorF("ERR [%s] %s: PANIC: %v", certSummary(connClientCert(qc)), qc.RemoteAddr(), r)
 			qc.Close()
 		}
 	}()
@@ -130,12 +129,12 @@ func (p *Proxy) handleQUICConnection(qc *netw.QUICConn) {
 	numOpen := p.inConns.add(qc)
 	qc.OnClose(func() {
 		p.inConns.remove(qc)
-		startTime := qc.Annotation(startTimeKey, time.Time{}).(time.Time)
-		log.Printf("END %s; Dur:%s Recv:%d Sent:%d",
-			formatConnDesc(qc), time.Since(startTime).Truncate(time.Millisecond),
-			qc.BytesReceived(), qc.BytesSent())
 		if be := connBackend(qc); be != nil {
 			be.incInFlight(-1)
+			startTime := qc.Annotation(startTimeKey, time.Time{}).(time.Time)
+			be.logConnF("END %s; Dur:%s Recv:%d Sent:%d",
+				formatConnDesc(qc), time.Since(startTime).Truncate(time.Millisecond),
+				qc.BytesReceived(), qc.BytesSent())
 		}
 		p.connClosed.Broadcast()
 	})
@@ -159,7 +158,7 @@ func (p *Proxy) handleQUICConnection(qc *netw.QUICConn) {
 	be, err := p.backend(cs.ServerName, cs.NegotiatedProtocol)
 	if err != nil {
 		p.recordEvent(err.Error())
-		log.Printf("BAD [%s] %s:%s ➔ %q: %v", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), cs.ServerName, err)
+		be.logErrorF("BAD [%s] %s:%s ➔ %q: %v", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), cs.ServerName, err)
 		qc.CloseWithError(quicUnrecognizedName, "unrecognized name")
 		return
 	}
@@ -169,7 +168,7 @@ func (p *Proxy) handleQUICConnection(qc *netw.QUICConn) {
 
 	if numOpen >= p.cfg.MaxOpen {
 		p.recordEvent("too many open connections")
-		log.Printf("ERR [%s] %s:%s: too many open connections: %d >= %d", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), numOpen, p.cfg.MaxOpen)
+		be.logErrorF("ERR [%s] %s:%s: too many open connections: %d >= %d", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), numOpen, p.cfg.MaxOpen)
 		return
 	}
 
@@ -179,16 +178,16 @@ func (p *Proxy) handleQUICConnection(qc *netw.QUICConn) {
 
 	if err := be.checkIP(qc.RemoteAddr()); err != nil {
 		p.recordEvent(idnaToUnicode(cs.ServerName) + " CheckIP " + err.Error())
-		log.Printf("BAD [%s] %s:%s ➔ %q CheckIP: %v", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), idnaToUnicode(cs.ServerName), err)
+		be.logErrorF("BAD [%s] %s:%s ➔ %q CheckIP: %v", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), idnaToUnicode(cs.ServerName), err)
 		qc.CloseWithError(quicAccessDenied, "access denied")
 		return
 	}
 
-	log.Printf("QUC [%s] %s:%s ➔ %s|%s:%s", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), idnaToUnicode(cs.ServerName), be.Mode, cs.NegotiatedProtocol)
+	be.logConnF("QUC [%s] %s:%s ➔ %s|%s:%s", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), idnaToUnicode(cs.ServerName), be.Mode, cs.NegotiatedProtocol)
 	if err := be.connLimit.Wait(ctx); err != nil {
 		if !errors.Is(err, context.Canceled) {
 			p.recordEvent(err.Error())
-			log.Printf("ERR [%s] %s ➔  %q Wait: %v", sum, qc.RemoteAddr(), idnaToUnicode(cs.ServerName), err)
+			be.logErrorF("ERR [%s] %s ➔  %q Wait: %v", sum, qc.RemoteAddr(), idnaToUnicode(cs.ServerName), err)
 		}
 		return
 	}
@@ -205,7 +204,7 @@ func (p *Proxy) handleQUICConnection(qc *netw.QUICConn) {
 		if errors.As(err, &idleTimeout) && idleTimeout.Timeout() {
 			return
 		}
-		log.Printf("ERR [%s] %s:%s ➔ %s|%s:%s %s: %v", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), idnaToUnicode(cs.ServerName), be.Mode, cs.NegotiatedProtocol, tag, err)
+		be.logErrorF("ERR [%s] %s:%s ➔ %s|%s:%s %s: %v", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), idnaToUnicode(cs.ServerName), be.Mode, cs.NegotiatedProtocol, tag, err)
 	}
 
 	if serv, ok := be.http3Server.(*http3.Server); ok && cs.NegotiatedProtocol == "h3" {
@@ -219,7 +218,7 @@ func (p *Proxy) handleQUICConnection(qc *netw.QUICConn) {
 		beConn, err := be.dialQUICBackend(ctx, cs.NegotiatedProtocol)
 		if err != nil {
 			qc.CloseWithError(quicBadGateway, "bad gateway")
-			log.Printf("ERR [%s] %s:%s ➔ %s|%s:%s dialQUICBackend: %v", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), idnaToUnicode(cs.ServerName), be.Mode, cs.NegotiatedProtocol, err)
+			be.logErrorF("ERR [%s] %s:%s ➔ %s|%s:%s dialQUICBackend: %v", sum, qc.RemoteAddr().Network(), qc.RemoteAddr(), idnaToUnicode(cs.ServerName), be.Mode, cs.NegotiatedProtocol, err)
 			return
 		}
 		defer beConn.Close()
@@ -370,7 +369,7 @@ func (p *Proxy) handleQUICTCPStream(ctx context.Context, be *Backend, conn *netw
 
 	switch be.Mode {
 	case ModeConsole, ModeLocal, ModeHTTP, ModeHTTPS:
-		log.Printf("STR %s", formatConnDesc(conn))
+		be.logConnF("STR %s", formatConnDesc(conn))
 		closeConnNeeded = false
 		be.httpConnChan <- conn
 
@@ -378,7 +377,7 @@ func (p *Proxy) handleQUICTCPStream(ctx context.Context, be *Backend, conn *netw
 		intConn, err := be.dial(ctx, connProto(conn))
 		if err != nil {
 			p.recordEvent("dial error")
-			log.Printf("ERR [-] %s:%s ➔  %q Dial: %v", conn.RemoteAddr().Network(), conn.RemoteAddr(), serverName, err)
+			be.logErrorF("ERR [-] %s:%s ➔  %q Dial: %v", conn.RemoteAddr().Network(), conn.RemoteAddr(), serverName, err)
 			return
 		}
 		defer intConn.Close()
@@ -390,22 +389,22 @@ func (p *Proxy) handleQUICTCPStream(ctx context.Context, be *Backend, conn *netw
 		}); ok {
 			cc.SetBridgeAddr(intConn.RemoteAddr().Network() + ":" + intConn.RemoteAddr().String())
 		}
-		log.Printf("STR %s", formatConnDesc(conn))
+		be.logConnF("STR %s", formatConnDesc(conn))
 
 		if err := be.bridgeConns(conn, intConn); err != nil {
-			log.Printf("DBG %s %v", formatConnDesc(conn), err)
+			be.logErrorF("DBG %s %v", formatConnDesc(conn), err)
 		}
 
 		startTime := conn.Annotation(startTimeKey, time.Time{}).(time.Time)
 		dialTime := conn.Annotation(dialDoneKey, time.Time{}).(time.Time)
 		totalTime := time.Since(startTime).Truncate(time.Millisecond)
 
-		log.Printf("END %s; Dial:%s Dur:%s Recv:%d Sent:%d", formatConnDesc(conn),
+		be.logConnF("END %s; Dial:%s Dur:%s Recv:%d Sent:%d", formatConnDesc(conn),
 			dialTime.Sub(startTime).Truncate(time.Millisecond), totalTime,
 			conn.BytesReceived(), conn.BytesSent())
 
 	default:
-		log.Printf("ERR [-] %s:%s: unhandled stream %q", conn.RemoteAddr().Network(), conn.RemoteAddr(), be.Mode)
+		be.logErrorF("ERR [-] %s:%s: unhandled stream %q", conn.RemoteAddr().Network(), conn.RemoteAddr(), be.Mode)
 	}
 }
 
@@ -413,7 +412,7 @@ func (be *Backend) handleQUICQUICStream(ctx context.Context, dest *netw.QUICConn
 	serverName := idnaToUnicode(connServerName(conn))
 	qs, ok := conn.Conn.(*netw.QUICStream)
 	if !ok {
-		log.Printf("ERR [-] %s ➔  %q not a QUICStream", conn.RemoteAddr(), serverName)
+		be.logErrorF("ERR [-] %s ➔  %q not a QUICStream", conn.RemoteAddr(), serverName)
 		return
 	}
 	var intConn *netw.Conn
@@ -421,7 +420,7 @@ func (be *Backend) handleQUICQUICStream(ctx context.Context, dest *netw.QUICConn
 		sendStream, err := dest.OpenUniStreamSync(ctx)
 		if err != nil {
 			be.recordEvent("openstream error")
-			log.Printf("ERR [-] %s:%s ➔  %q OpenUniStreamSync: %v", conn.RemoteAddr().Network(), conn.RemoteAddr(), serverName, err)
+			be.logErrorF("ERR [-] %s:%s ➔  %q OpenUniStreamSync: %v", conn.RemoteAddr().Network(), conn.RemoteAddr(), serverName, err)
 			var appErr *quic.ApplicationError
 			if errors.As(err, &appErr) {
 				qs.CancelRead(quic.StreamErrorCode(appErr.ErrorCode))
@@ -433,7 +432,7 @@ func (be *Backend) handleQUICQUICStream(ctx context.Context, dest *netw.QUICConn
 		stream, err := dest.OpenStreamSync(ctx)
 		if err != nil {
 			be.recordEvent("openstream error")
-			log.Printf("ERR [-] %s:%s ➔  %q OpenStreamSync: %v", conn.RemoteAddr().Network(), conn.RemoteAddr(), serverName, err)
+			be.logErrorF("ERR [-] %s:%s ➔  %q OpenStreamSync: %v", conn.RemoteAddr().Network(), conn.RemoteAddr(), serverName, err)
 			var appErr *quic.ApplicationError
 			if errors.As(err, &appErr) {
 				qs.CancelRead(quic.StreamErrorCode(appErr.ErrorCode))
@@ -449,17 +448,17 @@ func (be *Backend) handleQUICQUICStream(ctx context.Context, dest *netw.QUICConn
 	conn.SetAnnotation(dialDoneKey, now)
 
 	desc := formatConnDesc(conn)
-	log.Printf("STR %s", desc)
+	be.logConnF("STR %s", desc)
 
 	if err := be.bridgeConns(conn, intConn); err != nil {
-		log.Printf("DBG %s %v", desc, err)
+		be.logErrorF("DBG %s %v", desc, err)
 	}
 
 	startTime := conn.Annotation(startTimeKey, time.Time{}).(time.Time)
 	dialTime := conn.Annotation(dialDoneKey, time.Time{}).(time.Time)
 	totalTime := time.Since(startTime).Truncate(time.Millisecond)
 
-	log.Printf("END %s; Dial:%s Dur:%s Recv:%d Sent:%d", desc,
+	be.logConnF("END %s; Dial:%s Dur:%s Recv:%d Sent:%d", desc,
 		dialTime.Sub(startTime).Truncate(time.Millisecond), totalTime,
 		conn.BytesReceived(), conn.BytesSent())
 }
@@ -560,7 +559,7 @@ func (be *Backend) dialQUICBackend(ctx context.Context, proto string) (*netw.QUI
 		cancel()
 		if err != nil {
 			if max--; max > 0 {
-				log.Printf("ERR dialQUIC %q: %v", addr, err)
+				be.logErrorF("ERR dialQUIC %q: %v", addr, err)
 				continue
 			}
 			return nil, err