+++ proxy/revoke.go
@@ -31,7 +31,6 @@ import (
 	"encoding/pem"
 	"errors"
 	"fmt"
-	"log"
 	"sort"
 	"strings"
 	"time"
@@ -72,33 +71,33 @@ func (p *Proxy) RevokeAllCertificates(ctx context.Context, reason string) error 
 		Key:          accountKey,
 		UserAgent:    "tlsproxy",
 	}
-	log.Print("!!!")
-	log.Print("!!! WARNING")
-	log.Print("!!!")
+	p.logErrorF("!!!")
+	p.logErrorF("!!! WARNING")
+	p.logErrorF("!!!")
 	if n := len(toRevoke); n == 1 {
-		log.Print("!!! About to REVOKE 1 certificate:")
+		p.logErrorF("!!! About to REVOKE 1 certificate:")
 	} else {
-		log.Printf("!!! About to REVOKE %d certificates:", n)
+		p.logErrorF("!!! About to REVOKE %d certificates:", n)
 	}
-	log.Print("!!!")
+	p.logErrorF("!!!")
 	for _, key := range toRevoke {
-		log.Printf("!!!   %s", key)
+		p.logErrorF("!!!   %s", key)
 	}
-	log.Print("!!!")
-	log.Print("!!! Press CTRL-C now to abort.")
-	log.Print("!!!")
+	p.logErrorF("!!!")
+	p.logErrorF("!!! Press CTRL-C now to abort.")
+	p.logErrorF("!!!")
 	time.Sleep(10 * time.Second)
 
 	now := time.Now()
 	for _, key := range toRevoke {
 		if now.After(certs[key].Leaf.NotAfter) {
-			log.Printf("!!! Expired: %s", key)
+			p.logErrorF("!!! Expired: %s", key)
 			continue
 		}
 		if err := client.RevokeCert(ctx, certs[key].PrivateKey.(crypto.Signer), certs[key].Certificate[0], reasonCode); err != nil {
 			return err
 		}
-		log.Printf("!!! Revoked: %s", key)
+		p.logErrorF("!!! Revoked: %s", key)
 	}
 	return p.certManager.(*autocert.Manager).Cache.(*autocertcache.Cache).DeleteKeys(ctx, toRevoke)
 }
@@ -128,7 +127,7 @@ L:
 		}
 		if len(cert.Leaf.DNSNames) > 0 {
 			n := idnaToUnicode(cert.Leaf.DNSNames[0])
-			log.Printf("INF Unused certificate: %s (%s)", n, k)
+			p.logErrorF("INF Unused certificate: %s (%s)", n, k)
 			toRevoke = append(toRevoke, k)
 		}
 	}
@@ -136,7 +135,7 @@ L:
 
 	if !actuallyRevoke {
 		if len(toRevoke) > 0 {
-			log.Print("INF Set \"revokeUnusedCertificates: true\" to automatically revoke unused certificates")
+			p.logErrorF("INF Set \"revokeUnusedCertificates: true\" to automatically revoke unused certificates")
 		}
 		return nil
 	}
@@ -153,13 +152,13 @@ L:
 	now := time.Now()
 	for _, key := range toRevoke {
 		if now.After(certs[key].Leaf.NotAfter) {
-			log.Printf("INF Expired certificate: %s", key)
+			p.logErrorF("INF Expired certificate: %s", key)
 			continue
 		}
 		if err := client.RevokeCert(ctx, certs[key].PrivateKey.(crypto.Signer), certs[key].Certificate[0], acme.CRLReasonUnspecified); err != nil {
 			return err
 		}
-		log.Printf("INF Revoked unused certificate: %s", key)
+		p.logErrorF("INF Revoked unused certificate: %s", key)
 	}
 	return p.certManager.(*autocert.Manager).Cache.(*autocertcache.Cache).DeleteKeys(ctx, toRevoke)
 }
@@ -209,7 +208,7 @@ L:
 		}
 		data, err := cache.Get(ctx, k)
 		if err != nil {
-			log.Printf("ERR %s: %v", k, err)
+			p.logErrorF("ERR %s: %v", k, err)
 			continue
 		}
 		var privKey crypto.Signer
@@ -222,7 +221,7 @@ L:
 			if strings.Contains(b.Type, "PRIVATE KEY") {
 				var err error
 				if privKey, err = parsePrivateKey(b.Bytes); err != nil {
-					log.Printf("ERR %s: %v", k, err)
+					p.logErrorF("ERR %s: %v", k, err)
 					continue L
 				}
 			}
@@ -231,16 +230,16 @@ L:
 			}
 		}
 		if privKey == nil {
-			log.Printf("ERR %s: missing private key", k)
+			p.logErrorF("ERR %s: missing private key", k)
 			continue
 		}
 		if len(certs) == 0 {
-			log.Printf("ERR %s: missing cert", k)
+			p.logErrorF("ERR %s: missing cert", k)
 			continue
 		}
 		leaf, err := x509.ParseCertificate(certs[0])
 		if err != nil {
-			log.Printf("ERR %s: %s", k, err)
+			p.logErrorF("ERR %s: %s", k, err)
 			continue
 		}
 		out[k] = &tls.Certificate{